Make Trilinos Solver classes compatible with LinearOperators.
This patch adds some core functionality to the Trilinos Solver that is
later necessary for them to be compatible with LinearOperators.
Specifically, this is two additional solve() functions that accept
Trilinos EpetraOperators as the input Matrix (with one accomodating
wrapped Trilinos preconditioners, and the other another EpetraOperator
as a preconditioner.).